lessson-5.1
moved todo functionality into TodoContainer

added DELETE method to delete todo's

added a sort method from most recent to oldest
